Fix #4
Moved away from trying to parse number of seconds into a time object
since that causes a lot of issues especially with DST which I discovered
today to be one of the most disturbing ideas in the history of mankind.
Since I use these only for logging time durations, I added a new
dotoo#time#log for this purpose, which just takes todays date and
appends the hours/minutes to it and builds an object, since we are only
interested in the hours:mins this works.