Improved agenda view. Fix #99.
Introducing a new option `g:dotoo#agenda_views#agenda#start_of` that
controls the starting date for the agenda. If set to `span` (default) it
starts from the beginning of the span till the end of the span, whereas,
if set to `today` it starts from today's date till the number of days
the current span.