Fix #102. Fixes captured todo refiling on closing.
Due to a recent change in vim autocmds, it is no longer allowed to
create splits within `BufHidden` and related commands that run just
before closing / deleting / wiping a buffer.

In order to work around this limitation, capture now relies on `edit` to
open the capture target file to save the captured todo in it.