Allow clocking for habits
Headlines setup as a habit did not previously allow clocking time,
I believe it makes sense to have this feature, along with state changes
we should also be able to log time spent on these tasks and then have
them reported in the agenda view log summary. This commit fixes this.