Move to native `readfile()`. Fix #127
Since we moved to using `writefile`, the idiomatic approach is to also
move to `readfile`.

Because of how `writefile` works, it doesn't appear to refresh the
buffer of the file written to automatically, hence older self
implementation of `s:readfile` was unable to read the latest file from
disk, it required to remove the usage of `noauto` which may have
unintended consequences and performance implications.

In my tests, readfile seems to be able to read the files well without
worrying about buffers and whether they are synced from disk.

However, this may be error prone for older versions of vim. At this
point I have not been able to assess which minimum version of vim this
depends on.