Added a Wiki Agenda View to List / Search through all notes
Previous Note Agenda View required the presence of :NOTE: tag, the wiki
is all content that has `notes` somewhere in the path, hence is able to
get more content. Searching is the most useful and powerful way to go
through your wiki and you get it out of the box with the agenda view.